requirements.txt
affine==2.4.0
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
branca==0.8.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cligj==0.7.2
colorama==0.4.6
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.7
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-yasg==1.21.11
folium==0.20.0
geopandas==1.1.1
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
Jinja2==3.1.6
joblib==1.5.2
kombu==5.5.4
MarkupSafe==3.0.3
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
rasterio==1.4.3
redis==7.0.0
requests==2.32.5
requests-oauthlib==2.0.0
scikit-learn==1.7.2
scipy==1.16.2
shapely==2.1.2
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
xyzservices==2025.4.0
